requirements.txt
tensorflow
pickle 
matplotlib
numpy
selenium
Pillow
requests